Reorder and refactor
