+++ src/datachain/data_storage/metastore.py
@@ -27,12 +27,12 @@ from sqlalchemy import (
 )
 from sqlalchemy.sql import func as f
 
+from datachain.catalog.dependency import DatasetDependencyNode
 from datachain.checkpoint import Checkpoint
 from datachain.data_storage import JobQueryType, JobStatus
 from datachain.data_storage.serializer import Serializable
 from datachain.dataset import (
     DatasetDependency,
-    DatasetDependencyNode,
     DatasetListRecord,
     DatasetListVersion,
     DatasetRecord,
@@ -1519,7 +1519,6 @@ class AbstractDBMetastore(AbstractMetastore):
             & (dd.c.source_dataset_version_id == version_id)
         )
 
-        # Recursive CTE
         cte = base_query.cte(name="dependency_tree", recursive=True)
 
         # Recursive case: dependencies of dependencies
@@ -1537,27 +1536,27 @@ class AbstractDBMetastore(AbstractMetastore):
         cte = cte.union(recursive_query)
 
         # Fetch all with full details
-        final_query = (
-            select(
-                n.c.name,
-                p.c.name,
-                cte.c.id,
-                cte.c.dataset_id,
-                cte.c.dataset_version_id,
-                d.c.name,
-                dv.c.version,
-                dv.c.created_at,
-                cte.c.source_dataset_id,
-                cte.c.source_dataset_version_id,
-                cte.c.depth,
-            )
-            .select_from(
-                cte.join(d, cte.c.dataset_id == d.c.id, isouter=True)
-                .join(dv, cte.c.dataset_version_id == dv.c.id, isouter=True)
-                .join(p, d.c.project_id == p.c.id, isouter=True)
-                .join(n, p.c.namespace_id == n.c.id, isouter=True)
-            )
-            .order_by(cte.c.depth, cte.c.id)
+        final_query = select(
+            n.c.name,
+            p.c.name,
+            cte.c.id,
+            cte.c.dataset_id,
+            cte.c.dataset_version_id,
+            d.c.name,
+            dv.c.version,
+            dv.c.created_at,
+            cte.c.source_dataset_id,
+            cte.c.source_dataset_version_id,
+            cte.c.depth,
+        ).select_from(
+            # Use outer joins to handle cases where dependent datasets have been
+            # physically deleted. This allows us to return dependency records with
+            # None values instead of silently omitting them, making broken
+            # dependencies visible to callers.
+            cte.join(d, cte.c.dataset_id == d.c.id, isouter=True)
+            .join(dv, cte.c.dataset_version_id == dv.c.id, isouter=True)
+            .join(p, d.c.project_id == p.c.id, isouter=True)
+            .join(n, p.c.namespace_id == n.c.id, isouter=True)
         )
 
         return [
