+++ src/datachain/dataset.py
@@ -20,7 +20,6 @@ LT = TypeVar("LT", bound="DatasetListRecord")
 V = TypeVar("V", bound="DatasetVersion")
 LV = TypeVar("LV", bound="DatasetListVersion")
 DD = TypeVar("DD", bound="DatasetDependency")
-DDN = TypeVar("DDN", bound="DatasetDependencyNode")
 
 DATASET_PREFIX = "ds://"
 QUERY_DATASET_PREFIX = "ds_query_"
@@ -93,62 +92,6 @@ def parse_dataset_name(name: str) -> tuple[str | None, str | None, str]:
 class DatasetDependencyType:
     DATASET = "dataset"
     STORAGE = "storage"
-
-
-@dataclass
-class DatasetDependencyNode:
-    namespace: str
-    project: str
-    id: int
-    dataset_id: int | None
-    dataset_version_id: int | None
-    dataset_name: str | None
-    dataset_version: str | None
-    created_at: datetime
-    source_dataset_id: int
-    source_dataset_version_id: int | None
-    depth: int
-
-    @classmethod
-    def parse(
-        cls: builtins.type[DDN],
-        namespace: str,
-        project: str,
-        id: int,
-        dataset_id: int | None,
-        dataset_version_id: int | None,
-        dataset_name: str | None,
-        dataset_version: str | None,
-        created_at: datetime,
-        source_dataset_id: int,
-        source_dataset_version_id: int | None,
-        depth: int,
-    ) -> "DatasetDependencyNode | None":
-        return cls(
-            namespace,
-            project,
-            id,
-            dataset_id,
-            dataset_version_id,
-            dataset_name,
-            dataset_version,
-            created_at,
-            source_dataset_id,
-            source_dataset_version_id,
-            depth,
-        )
-
-    def to_dependency(self) -> "DatasetDependency | None":
-        return DatasetDependency.parse(
-            namespace_name=self.namespace,
-            project_name=self.project,
-            id=self.id,
-            dataset_id=self.dataset_id,
-            dataset_version_id=self.dataset_version_id,
-            dataset_name=self.dataset_name,
-            dataset_version=self.dataset_version,
-            dataset_version_created_at=self.created_at,
-        )
 
 
 @dataclass
